Make the map work on safari + made the code more modular
