+++ client/src/Components/TemporaryHome/index.tsx
@@ -1,7 +1,6 @@
-import React, { useEffect, useRef, useState } from 'react'
+import React from 'react'
 import Map from 'Components/Map'
 import Dialogue from 'Components/Dialogue'
-import { Location } from 'lib/types'
 import bananamanAvatar from 'lib/assets/banana.jpg'
 import AudioButton from 'Components/AudioButton'
 
@@ -16,30 +15,10 @@ const messages = [
 ]
 
 const TemporaryHome: React.FC = () => {
-  const [height, setHeight] = useState(0)
-  const elementRef = useRef(null as null | HTMLDivElement)
-
-  useEffect(() => {
-    if (elementRef?.current?.clientHeight) {
-      setHeight(elementRef?.current?.clientHeight)
-    }
-  }, []) //empty dependency array so it only runs once at render
-
-  const [selected, onSelect] = useState<Location | null>(null)
-
   return (
     <>
       <AudioButton soundFile={songLocation} loop={true} />
-      <div className='flex-auto'>
-        <div
-          ref={elementRef}
-          className='h-full 2xl:h-4/5 xl:h-4/5 flex justify-center'
-        >
-          {height === 0 ? null : (
-            <Map selected={selected} onSelect={onSelect} height={height} />
-          )}
-        </div>
-      </div>
+      <Map />
       <Dialogue
         messages={messages}
         speaker={'bananaman'}
